refactor shell extension project
