--- crates/devolutions-pedm-contextmenu/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "devolutions-pedm-contextmenu"
-version = "0.0.0"
-edition = "2021"
-license = "MIT/Apache-2.0"
-authors = ["Devolutions Inc. <infos@devolutions.net>"]
-description = "Devolutions PEDM Windows shell extension package"
-publish = false
-
-[dependencies]
-devolutions-pedm-contextmenu-shell = { path = "./devolutions-pedm-contextmenu-shell" }
-
-[build-dependencies]
-devolutions-pedm-shared = { path = "../devolutions-pedm-shared", features = ["build"] }
-fs_extra = "1.3.0"
