--- crates/devolutions-pedm-contextmenu/devolutions-pedm-contextmenu-shell/resources.rc
@@ -1,1 +1,0 @@
-101 ICON "../../../package/AgentWindowsManaged/Resources/DevolutionsAgent.ico"
