--- crates/devolutions-pedm-contextmenu/src/main.rs
@@ -1,2 +1,0 @@
-// Empty because an executable is needed to bundle the Appx, however we only want a library for the context menu.
-fn main() {}
