+++ crates/devolutions-pedm-shell-ext/AppxManifest.xml
@@ -1,0 +1,79 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Package xmlns="http://schemas.microsoft.com/appx/manifest/foundation/windows10"
+    xmlns:uap="http://schemas.microsoft.com/appx/manifest/uap/windows10"
+    xmlns:uap2="http://schemas.microsoft.com/appx/manifest/uap/windows10/2"
+    xmlns:uap3="http://schemas.microsoft.com/appx/manifest/uap/windows10/3"
+    xmlns:rescap="http://schemas.microsoft.com/appx/manifest/foundation/windows10/restrictedcapabilities"
+    xmlns:desktop="http://schemas.microsoft.com/appx/manifest/desktop/windows10"
+    xmlns:desktop4="http://schemas.microsoft.com/appx/manifest/desktop/windows10/4"
+    xmlns:desktop5="http://schemas.microsoft.com/appx/manifest/desktop/windows10/5"
+    xmlns:uap10="http://schemas.microsoft.com/appx/manifest/uap/windows10/10"
+    xmlns:com="http://schemas.microsoft.com/appx/manifest/com/windows10"
+    IgnorableNamespaces="uap uap2 uap3 rescap desktop desktop4 desktop5 uap10 com">
+    <Identity Name="DevolutionsAgent" ProcessorArchitecture="neutral" Publisher="CN=Devolutions"
+        Version="1.0.0.0" />
+    <Properties>
+        <DisplayName>Devolutions Agent Shell Extension</DisplayName>
+        <PublisherDisplayName>Devolutions</PublisherDisplayName>
+        <Logo>Assets\Logo.png</Logo>
+        <uap10:AllowExternalContent>true</uap10:AllowExternalContent>
+    </Properties>
+    <Resources>
+        <Resource Language="en-us" />
+    </Resources>
+    <Dependencies>
+        <TargetDeviceFamily Name="Windows.Desktop"
+            MinVersion="10.0.19000.0" MaxVersionTested="10.0.19000.0" />
+    </Dependencies>
+    <Capabilities>
+        <rescap:Capability Name="runFullTrust" />
+        <rescap:Capability Name="unvirtualizedResources" />
+    </Capabilities>
+    <Applications>
+        <Application Id="DevolutionsAgent" Executable="DevolutionsAgent.exe"
+            uap10:TrustLevel="mediumIL" uap10:RuntimeBehavior="win32App">
+            <uap:VisualElements AppListEntry="none" DisplayName="Devolutions Agent shell extension"
+                Description="DevolutionsAgent" BackgroundColor="transparent"
+                Square150x150Logo="Assets\Square150x150Logo.png"
+                Square44x44Logo="Assets\Square44x44Logo.png">
+                <uap:DefaultTile Wide310x150Logo="Assets\Wide310x150Logo.png"
+                    Square310x310Logo="Assets\LargeTile.png" Square71x71Logo="Assets\SmallTile.png"></uap:DefaultTile>
+            </uap:VisualElements>
+            <Extensions>
+                <desktop4:Extension Category="windows.fileExplorerContextMenus">
+                    <desktop4:FileExplorerContextMenus>
+                        <desktop5:ItemType Type=".exe">
+                            <desktop5:Verb Id="Elevate"
+                                Clsid="0ba604fd-4a5a-4abb-92b1-09ac5c3bf356" />
+                        </desktop5:ItemType>
+                        <desktop5:ItemType Type=".msi">
+                            <desktop5:Verb Id="Elevate"
+                                Clsid="0ba604fd-4a5a-4abb-92b1-09ac5c3bf356" />
+                        </desktop5:ItemType>
+                        <desktop5:ItemType Type=".lnk">
+                            <desktop5:Verb Id="Elevate"
+                                Clsid="0ba604fd-4a5a-4abb-92b1-09ac5c3bf356" />
+                        </desktop5:ItemType>
+                        <desktop5:ItemType Type=".ps1">
+                            <desktop5:Verb Id="Elevate"
+                                Clsid="0ba604fd-4a5a-4abb-92b1-09ac5c3bf356" />
+                        </desktop5:ItemType>
+                        <desktop5:ItemType Type=".bat">
+                            <desktop5:Verb Id="Elevate"
+                                Clsid="0ba604fd-4a5a-4abb-92b1-09ac5c3bf356" />
+                        </desktop5:ItemType>
+                    </desktop4:FileExplorerContextMenus>
+                </desktop4:Extension>
+                <com:Extension Category="windows.comServer"
+                    uap10:RuntimeBehavior="packagedClassicApp">
+                    <com:ComServer>
+                        <com:SurrogateServer DisplayName="Devolutions Agent Shell Extension">
+                            <com:Class Id="0ba604fd-4a5a-4abb-92b1-09ac5c3bf356"
+                                Path="DevolutionsPedmShellExt.dll" ThreadingModel="STA" />
+                        </com:SurrogateServer>
+                    </com:ComServer>
+                </com:Extension>
+            </Extensions>
+        </Application>
+    </Applications>
+</Package>
