+++ crates/devolutions-pedm-shell-ext/Assets/LockScreenLogo.scale-200.png
