+++ crates/devolutions-pedm-shell-ext/Assets/SplashScreen.scale-200.png
