+++ crates/devolutions-pedm-shell-ext/Assets/Square150x150Logo.scale-200.png
