+++ crates/devolutions-pedm-shell-ext/Assets/Square44x44Logo.scale-200.png
