+++ crates/devolutions-pedm-shell-ext/Assets/StoreLogo.png
