+++ crates/devolutions-pedm-shell-ext/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "devolutions-pedm-contextmenu-shell"
+name = "devolutions-pedm-shell-ext"
 version = "0.0.0"
 edition = "2021"
 license = "MIT/Apache-2.0"
@@ -13,9 +13,13 @@ crate-type = ["cdylib"]
 [dependencies]
 tokio = { version = "1.38.0", features = ["sync"] }
 windows-core = "0.58.0" # necessary because of macro resolving `windows::core`. Must match `win-api-wrappers` version.
-devolutions-pedm-shared = { path = "../../devolutions-pedm-shared", features = ["pedm_client", "desktop"] }
-win-api-wrappers = { path = "../../win-api-wrappers" }
+devolutions-pedm-shared = { path = "../devolutions-pedm-shared", features = ["pedm_client", "desktop"] }
+win-api-wrappers = { path = "../win-api-wrappers" }
 parking_lot = "0.12.3"
+
+[build-dependencies]
+devolutions-pedm-shared = { path = "../devolutions-pedm-shared", features = ["build"] }
+fs_extra = "1.3.0"
 
 [target.'cfg(windows)'.build-dependencies]
 embed-resource = "2.4"
