+++ crates/devolutions-pedm-shell-ext/resources.rc
@@ -1,0 +1,1 @@
+101 ICON "../../package/AgentWindowsManaged/Resources/DevolutionsAgent.ico"
